+++ R/coxph.R
@@ -97,7 +97,7 @@ coxModule <- function(input, output, session, data, data_label, data_varStruct =
     class01_factor <- unlist(data()[, lapply(.SD, function(x){identical(levels(x), c("0", "1"))}), .SDcols = factor_vars])
 
     validate(
-      need(!is.null(class01_factor), "No categorical variables coded as 0, 1 in data")
+      need(length(class01_factor) >=1, "No categorical variables coded as 0, 1 in data")
     )
     factor_01vars <- factor_vars[class01_factor]
 
@@ -115,10 +115,11 @@ coxModule <- function(input, output, session, data, data_label, data_varStruct =
   })
 
   output$eventtime <- renderUI({
-    req(!is.null(vlist()$factor_01vars))
     validate(
-      need(!is.null(vlist()$conti_list), "No continuous variables to be time")
+      need(length(vlist()$factor_01vars) >=1 , "No candidate event variables coded as 0, 1"),
+      need(length(vlist()$conti_list) >=1, "No candidate time variables")
     )
+
 
     tagList(
       selectInput(session$ns("event_cox"), "Event",
@@ -237,6 +238,8 @@ coxModule <- function(input, output, session, data, data_label, data_varStruct =
 
 
   out <- reactive({
+    req(!is.null(input$event_km))
+    req(!is.null(input$time_km))
     data.cox <- data()
     data.cox[[input$event_cox]] <- as.numeric(as.vector(data.cox[[input$event_cox]]))
     if(input$subcheck == T){